Move controlelr utils to utils
