+++ actions/list-controllers
@@ -8,7 +8,7 @@ from charms.layer.basic import activate_venv  # noqa: E402
 activate_venv()
 
 from charmhelpers.core import hookenv  # noqa: E402
-from controller.helpers import get_controllers  # noqa: E402
+from utils import get_controllers  # noqa: E402
 
 
 def list_controllers():
